+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - MeCard and vCard data type support.
-- Image merge support for WebP and SVG formats.
+- Image merge support for WebP, SVG and EPS image formats.
 - GD Backend support for PNG and WebP generation (useful when Imagick is not available).
 
 ### Changed